update Device.GRE.Tunnel.1.DefaultDSCPMark data type integer -> unsigned integer ( OFW-2081 )
Fix SPV on WiFi params fail with FaultString "Invalid Parameter Value"

BUG DESCRIPTION
Due to wrong data type for  parameter, SPV was not working

FIX DESCRIPTION
Data type was changed from integer to unsigned integer.

TESTING PERFORMED
console testing
dmcli eRT setv Device.GRE.Tunnel.1.DefaultDSCPMark uint 32
dmcli eRT getv Device.GRE.Tunnel.1.DefaultDSCPMark
checked in ACS as well

+++ custom/comcast/source/TR-181/custom_include/cosa_x_comcast-com_gre_apis.h
@@ -184,7 +184,7 @@ _COSA_DML_GRE_TUNNEL
     BOOL                        UseSequenceNumber;
     BOOL                        UseChecksum;
     
-	INT                         DSCPMarkPolicy;
+	ULONG                       DSCPMarkPolicy;
 	INT                         GreTransportInterface;
 	
     COSA_DML_KEEPALIVE_POLICY   KeepAlivePolicy;
@@ -322,10 +322,10 @@ ANSC_STATUS
 CosaDml_GreTunnelSetUseChecksum(ULONG ins, BOOL enable);	//
 
 ANSC_STATUS
-CosaDml_GreTunnelGetDSCPMarkPolicy(ULONG ins, INT *dscp); //
+CosaDml_GreTunnelGetDSCPMarkPolicy(ULONG ins, ULONG *dscp); //
 
 ANSC_STATUS
-CosaDml_GreTunnelSetDSCPMarkPolicy(ULONG ins, INT dscp); //
+CosaDml_GreTunnelSetDSCPMarkPolicy(ULONG ins, ULONG dscp); //
 
 ANSC_STATUS
 CosaDml_GreTunnelIfGetVlanId(ULONG tuIns, ULONG ins, INT *vlanId); //
